use mockery for test mocks

+++ relayer/ocr2/types.go
@@ -33,8 +33,8 @@ type TransmissionDetails struct {
 }
 
 type BillingDetails struct {
-	ObservationPaymentGJuels  uint64
-	TransmissionPaymentGJuels uint64
+	ObservationPaymentGJuels  uint32
+	TransmissionPaymentGJuels uint32
 }
 
 type RoundData struct {
